requirements.txt
bandit
google-generativeai
requests
python-dotenv
